pyproject.toml
[tool.poetry]
name = "smit-test"
version = "0.1.0"
description = ""
authors = ["shmookoff <shmookoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
aerich = "^0.7.1"
pydantic = "^2"
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.4.1"


[tool.aerich]
tortoise_orm = "smit_test.settings.tortoise_config"
location = "./smit_test/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true